att: corrigindo estrutura
