+++ src/utils/functions/generateKey.ts
