+++ src/utils/functions/getTruncateString.ts
